requirements.txt
python-telegram-bot[job-queue]>=21.10
gTTS>=2.5.1
aiosqlite>=0.19.0
python-dotenv>=1.0.1
